requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
ultralytics==8.3.40
easyocr==1.7.2
requests==2.32.3
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
jinja2==3.1.4
python-multipart==0.0.17
streamlit==1.40.1
streamlit-autorefresh==1.0.1
pandas==2.2.3
python-dotenv==1.0.1
slowapi==0.1.9
psycopg2-binary==2.9.10
